Language.Haskell.Tools.Rewrite.Create.Binds

Description

Generation of binding-level AST fragments for refactorings. The bindings defined here create a the annotated version of the AST constructor with the same name. For example, mkMatch creates the annotated version of the UMatch constructor.

Documentation

mkSimpleBind' :: Name -> Expr -> ValueBind

A simplified function to generate simple value bindings without local definitions, guards or complex lhs.

mkSimpleBind :: Pattern -> Rhs -> Maybe LocalBinds -> ValueBind

Creates a value binding ( v = "12" ).

mkFunctionBind :: [Match] -> ValueBind

Creates a function binding ( f 0 = 1; f x = x ). All matches must have the same name.

mkFunctionBind' :: Name -> [([Pattern], Expr)] -> ValueBind

A simplified function for creating function bindings without local definitions or guards.

mkMatch :: MatchLhs -> Rhs -> Maybe LocalBinds -> Match

Creates a clause of function binding

mkMatchLhs :: Name -> [Pattern] -> MatchLhs

Creates a match lhs with the function name and parameter names ( f a b )

mkInfixLhs :: Pattern -> Operator -> Pattern -> [Pattern] -> MatchLhs

Creates an infix match lhs for an operator ( a + b )

mkLocalBinds :: [LocalBind] -> MaybeLocalBinds

Local bindings attached to a declaration ( where x = 42 )

mkLocalBinds' :: [LocalBind] -> LocalBinds

mkLocalValBind :: ValueBind -> LocalBind

Creates a local binding for a value

mkLocalTypeSig :: TypeSignature -> LocalBind

Creates a local type signature

mkLocalFixity :: FixitySignature -> LocalBind

Creates a local fixity declaration

mkTypeSignature :: Name -> Type -> TypeSignature

Creates a type signature ( f :: Int -> Int )

mkInfixL :: Int -> Operator -> FixitySignature

Creates a left-associative fixity declaration ( infixl 5 +, - ).

mkInfixR :: Int -> Operator -> FixitySignature

Creates a right-associative fixity declaration ( infixr 5 +, - ).

mkInfix :: Int -> Operator -> FixitySignature

Creates a non-associative fixity declaration ( infix 5 +, - ).

mkUnguardedRhs :: Expr -> Rhs

Creates an unguarded right-hand-side ( = 3 )

mkGuardedRhss :: [GuardedRhs] -> Rhs

Creates an unguarded right-hand-side ( | x == 1 = 3; | otherwise = 4 )

mkGuardedRhs :: [RhsGuard] -> Expr -> GuardedRhs

Creates a guarded right-hand side of a value binding ( | x > 3 = 2 )

mkGuardBind :: Pattern -> Expr -> RhsGuard

Creates a bind statement in a pattern guard ( Just v <- x )

mkGuardLet :: [LocalBind] -> RhsGuard

Creates a let statement in a pattern guard ( let x = 3 )

mkGuardCheck :: Expr -> RhsGuard

Creates an expression to check for a pattern guard