Language.Haskell.Tools.Rewrite.Create.Exprs

Description

Generation of expression-level AST fragments for refactorings. The bindings defined here create a the annotated version of the AST constructor with the same name. For example, mkApp creates the annotated version of the App AST constructor.

Expressions

mkVar :: Name -> Expr

Create a expression for a variable or a data constructor ( a )

mkLit :: Literal -> Expr

Create a literal expression ( 42 )

mkInfixApp :: Expr -> Operator -> Expr -> Expr

Create a infix operator application expression ( a + b )

mkPrefixApp :: Operator -> Expr -> Expr

Create a prefix operator application expression ( -x )

mkApp :: Expr -> Expr -> Expr

Create a function application expression ( f 4 )

mkLambda :: [Pattern] -> Expr -> Expr

Create a lambda expression ( \a b -> a + b )

mkLet :: [LocalBind] -> Expr -> Expr

Create a local binding ( let x = 2; y = 3 in e x y )

mkIf :: Expr -> Expr -> Expr -> Expr

Create a if expression ( if a then b else c )

mkMultiIf :: [GuardedCaseRhs] -> Expr

Create a multi way if expressions with MultiWayIf extension ( if | guard1 -> expr1; guard2 -> expr2 )

mkCase :: Expr -> [Alt] -> Expr

Create a pattern matching expression ( case expr of pat1 -> expr1; pat2 -> expr2 )

mkDoBlock :: [Stmt] -> Expr

Create a do-notation expressions ( do x <- act1; act2 )

mkMDoBlock :: [Stmt] -> Expr

Create a mdo-notation expressions ( mdo x <- act1; act2 )

mkTuple :: [Expr] -> Expr

Create a tuple expression ( (e1, e2, e3) )

mkUnboxedTuple :: [Expr] -> Expr

Create a unboxed tuple expression ( (# e1, e2, e3 #) )

mkTupleSection :: [Maybe Expr] -> Expr

Create a tuple section, enabled with TupleSections ( (a,,b) ). One of the elements must be missing.

mkTupleUnboxedSection :: [Maybe Expr] -> Expr

Create a unboxed tuple section, enabled with TupleSections ( (#a,,b#) ). One of the elements must be missing.

mkList :: [Expr] -> Expr

Create a list expression: [1,2,3]

mkParArray :: [Expr] -> Expr

Create a parallel array expression: [: 1,2,3 :]

mkParen :: Expr -> Expr

Create a parenthesized expression: ( a + b )

mkLeftSection :: Expr -> Operator -> Expr

Create a left operator section: (1+)

mkRightSection :: Operator -> Expr -> Expr

Create a right operator section: (+1)

mkRecCon :: Name -> [FieldUpdate] -> Expr

Create a record value construction: Point { x = 3, y = -2 }

mkRecUpdate :: Expr -> [FieldUpdate] -> Expr

Create a record value update: p1 { x = 3, y = -2 }

mkEnum :: Expr -> Maybe Expr -> Maybe Expr -> Expr

Create a enumeration expression ( [1,3..10] )

mkParArrayEnum :: Expr -> Maybe Expr -> Expr -> Expr

Create a parallel array enumeration ( [: 1,3 .. 10 :] )

mkListComp :: Expr -> [ListCompBody] -> Expr

Create a list comprehension ( [ (x, y) | x <- xs | y <- ys ] )

mkParArrayComp :: Expr -> [ListCompBody] -> Expr

Create a parallel array comprehensions [: (x, y) | x <- xs , y <- ys :] enabled by ParallelArrays

mkExprTypeSig :: Expr -> Type -> Expr

Create a explicit type signature ( x :: Int )

mkExplicitTypeApp :: Expr -> Type -> Expr

Create a explicit type application ( show @Integer (read "5") )

mkVarQuote :: Name -> Expr

'x for template haskell reifying of expressions

mkTypeQuote :: Name -> Expr

''T for template haskell reifying of types

mkBracketExpr :: Bracket -> Expr

Create a template haskell bracket expression

mkSpliceExpr :: Splice -> Expr

Create a template haskell splice expression, for example: $(gen a) or $x

mkQuasiQuoteExpr :: QuasiQuote -> Expr

Create a template haskell quasi quote expression, for example: [quoter| a + b ]

mkExprPragma :: ExprPragma -> Expr -> Expr

Creates a pragma that marks an expression.

mkProcExpr :: Pattern -> Cmd -> Expr

Create a arrow definition: proc a -> f -< a+1

mkArrowApp :: Expr -> ArrowApp -> Expr -> Expr

Create a arrow definition: proc a -> f -< a+1

mkLambdaCase :: [Alt] -> Expr

Create a lambda case ( case 0 -> 1; 1 -> 2 )

mkStaticPointer :: Expr -> Expr

Create a static pointer expression ( static e ). The inner expression must be closed (cannot have variables bound outside)

Field updates

mkFieldUpdate :: Name -> Expr -> FieldUpdate

Create a update of a field ( x = 1 )

mkFieldPun :: Name -> FieldUpdate

Create a update the field to the value of the same name ( x )

mkFieldWildcard :: FieldUpdate

Create a update the fields of the bounded names to their values ( .. ). Must be the last initializer. Cannot be used in a record update expression.

Pattern matching and guards

mkAlt :: Pattern -> CaseRhs -> Maybe LocalBinds -> Alt

Create a clause of case expression ( Just x -> x + 1 )

mkCaseRhs :: Expr -> CaseRhs

Create a unguarded right-hand side a pattern match ( -> 3 )

mkGuardedCaseRhss :: [GuardedCaseRhs] -> CaseRhs

Create a guarded right-hand sides of a pattern match ( | x == 1 -> 3; | otherwise -> 4 )

mkGuardedCaseRhs :: [RhsGuard] -> Expr -> GuardedCaseRhs

Creates a guarded right-hand side of pattern matches binding ( | x > 3 -> 2 )

Pragmas that can be applied to expressions

mkCorePragma :: String -> ExprPragma

Creates a CORE pragma for adding notes to expressions.

mkSccPragma :: String -> ExprPragma

Creates an SCC pragma for defining cost centers for profiling

mkGeneratedPragma :: SourceRange -> ExprPragma

Creates a pragma that describes if an expression was generated from a code fragment by an external tool ( {-# GENERATED "Happy.y" 1:15-1:25 #-} )

mkSourceRange :: String -> Integer -> Integer -> Integer -> Integer -> SourceRange

Create a in-AST source ranges (for generated pragmas)

Commands

mkArrowAppCmd :: Expr -> ArrowApp -> Expr -> Cmd

An arrow application command ( f -< x + 1 )

mkArrowFromCmd :: Expr -> [Cmd] -> Cmd

A form command ( (|untilA (increment -< x+y) (within 0.5 -< x)|) )

mkAppCmd :: Cmd -> Expr -> Cmd

A function application command

mkInfixCmd :: Cmd -> Name -> Cmd -> Cmd

An infix command application

mkLambdaCmd :: [Pattern] -> Cmd -> Cmd

A lambda command

mkParenCmd :: Cmd -> Cmd

A parenthesized command

mkCaseCmd :: Expr -> [CmdAlt] -> Cmd

A pattern match command

mkIfCmd :: Expr -> Cmd -> Cmd -> Cmd

An if command ( if f x y then g -< x+1 else h -< y+2 )

mkLetCmd :: [LocalBind] -> Cmd -> Cmd

A local binding command ( let z = x+y )

mkDoCmd :: [CmdStmt] -> Cmd

A do-notation in a command

mkLeftAppl :: ArrowApp

Left arrow application: -<

mkRightAppl :: ArrowApp

Right arrow application: >-

mkLeftHighAppl :: ArrowApp

Left arrow high application: -<<

mkRightHighAppl :: ArrowApp

Right arrow high application: >>-

mkHole :: Expr

A hole expression _