Language.Haskell.Tools.Rewrite.Create.Types

Description

Generation of type-level AST fragments for refactorings. The bindings defined here create a the annotated version of the AST constructor with the same name. For example, mkTyForall creates the annotated version of the TyForall AST constructor.

Generation of types

mkForallType :: [TyVar] -> Type -> Type

Forall types ( forall x y . type )

mkTypeVar' :: Name -> TyVar

Simplified creation of type variables

mkCtxType :: Context -> Type -> Type

Type with a context ( forall x y . type )

mkFunctionType :: Type -> Type -> Type

Function types ( a -> b )

mkTupleType :: [Type] -> Type

Tuple types ( (a,b) )

mkUnboxedTupleType :: [Type] -> Type

Unboxed tuple types ( (#a,b#) )

mkListType :: Type -> Type

List type with special syntax ( [a] )

mkParArrayType :: Type -> Type

Parallel array type ( [:a:] )

mkTypeApp :: Type -> Type -> Type

Type application ( F a )

mkInfixTypeApp :: Type -> Operator -> Type -> Type

Infix type constructor ( (a <: b) )

mkParenType :: Type -> Type

Type surrounded by parentheses ( (T a) )

mkTypeVar :: Name -> TyVar

Creates a simple type variable

mkKindedTypeVar :: Name -> Kind -> TyVar

Creates a type variable with kind specification ( t :: * )

mkVarType :: Name -> Type

Type variable or constructor ( a )

mkKindedType :: Type -> Kind -> Type

Type with explicit kind signature ( a :: * )

mkBangType :: Type -> Type

Strict type marked with !.

mkLazyType :: Type -> Type

Lazy type marked with ~. (Should only be used if Strict or StrictData language extension is used)

mkUnpackType :: Type -> Type

Strict type marked with UNPACK pragma. (Usually contains the bang mark.)

mkNoUnpackType :: Type -> Type

Strict type marked with UNPACK pragma. (Usually contains the bang mark.)

mkWildcardType :: Type

A wildcard type ( _ ) with -XPartialTypeSignatures

mkNamedWildcardType :: Name -> Type

A named wildcard type ( _t ) with -XPartialTypeSignatures

mkSpliceType :: Splice -> Type

A Template Haskell splice type ( $(genType) ).

mkQuasiQuoteType :: QuasiQuote -> Type

A Template Haskell quasi-quote type ( [quoter| ... ] ).

mkPromotedIntType :: Integer -> Type

Numeric value promoted to the kind level.

mkPromotedStringType :: String -> Type

String value promoted to the kind level.

mkPromotedConType :: Name -> Type

A data constructor value promoted to the kind level.

mkPromotedListType :: [Type] -> Type

A list of elements as a kind.

mkPromotedTupleType :: [Type] -> Type

A tuple of elements as a kind.

mkPromotedUnitType :: Type

Kind of the unit value ().

Generation of contexts

mkContext :: Assertion -> Context

Creates a context of assertions ( C a => ... )

Generation of assertions

mkClassAssert :: Name -> [Type] -> Assertion

Class assertion (Cls x)

mkInfixAssert :: Type -> Operator -> Type -> Assertion

Infix class assertion, also contains type equations ( a ~ X y )

mkImplicitAssert :: Name -> Type -> Assertion

Creates an assertion for implicit parameter binding ( ?cmp :: a -> a -> Bool )

mkTupleAssertion :: [Assertion] -> Assertion

Creates a list of assertions ( (Eq a, Show a) )