Language.Haskell.Tools.Rewrite.Create.Modules

Description

Generation of UModule-level AST fragments for refactorings. The bindings defined here create a the annotated version of the AST constructor with the same name. For example, mkModule creates the annotated version of the UModule AST constructor.

Documentation

mkModule :: [FilePragma] -> Maybe ModuleHead -> [ImportDecl] -> [Decl] -> Module

The representation of a haskell module, that is a separate compilation unit. It may or may not have a header.

mkModuleHead :: ModuleName -> Maybe ModulePragma -> Maybe ExportSpecs -> ModuleHead

Module declaration with name and (optional) exports

mkExportSpecs :: [ExportSpec] -> ExportSpecs

A list of export specifications surrounded by parentheses

mkExportSpec :: IESpec -> ExportSpec

Export a name and related names

mkModuleExport :: ModuleName -> ExportSpec

The export of an imported module ( module A )

mkIESpec :: Name -> Maybe SubSpec -> IESpec

Marks a name to be imported or exported with related names (subspecifier)

mkPatternIESpec :: Name -> IESpec

Marks a pattern synonym to be imported or exported

mkSubList :: [Name] -> SubSpec

(a,b,c): a class exported with some of its methods, or a datatype exported with some of its constructors.

mkSubAll :: SubSpec

(..): a class exported with all of its methods, or a datatype exported with all of its constructors.

mkImportDecl :: Bool -> Bool -> Bool -> Maybe String -> ModuleName -> Maybe ModuleName -> Maybe ImportSpec -> ImportDecl

An import declaration: import Module.Name

mkImportSpecList :: [IESpec] -> ImportSpec

Restrict the import definition to ONLY import the listed names

mkImportHidingList :: [IESpec] -> ImportSpec

Restrict the import definition to DONT import the listed names

mkModuleName :: String -> ModuleName

The name of a module

Pragmas

mkFilePragmas :: [FilePragma] -> FilePragmaList

mkLanguagePragma :: [String] -> FilePragma

LANGUAGE pragma, listing the enabled language extensions in that file

mkOptionsGHC :: String -> FilePragma

OPTIONS pragma, possibly qualified with a tool, e.g. OPTIONS_GHC

mkModuleWarningPragma :: [String] -> ModulePragma

A warning pragma attached to the module

mkModuleDeprecatedPragma :: [String] -> ModulePragma

A deprecated pragma attached to the module